Removed all the server related code from renderer.rs and moved it to server.rs Renamed setup back to start
