+++ examples/maze/main.rs
@@ -17,7 +17,7 @@ const WIDTH: f64 = 600.0; // px
 const HEIGHT: f64 = 600.0; // px
 
 fn main() {
-    turtle::setup();
+    turtle::start();
 
     let maze = Maze::generate();
