+++ src/lib.rs
@@ -56,7 +56,7 @@ pub mod color;
 pub mod event;
 pub mod rand;
 
-pub use renderer::setup;
+pub use server::start;
 pub use turtle::{Turtle, Point, Distance, Angle};
 pub use speed::{Speed};
 pub use color::{Color};
